Feat #72 add Presenter-Container
